# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Babis M1 AI Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatSession {
  id: number;
  title: string;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ChatSessionInput {
  title?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface Message {
  id: number;
  sessionId: number;
  role: MessageRole;
  content: string;
  thinkingMode?: boolean;
  createdAt: string;
}

export interface MessageInput {
  content: string;
  thinkingMode?: boolean;
}

export type TrainingStatusStatus = typeof TrainingStatusStatus[keyof typeof TrainingStatusStatus];


export const TrainingStatusStatus = {
  idle: 'idle',
  initializing: 'initializing',
  running: 'running',
  paused: 'paused',
  stopped: 'stopped',
  error: 'error',
} as const;

export type TrainingStatusPowerMode = typeof TrainingStatusPowerMode[keyof typeof TrainingStatusPowerMode];


export const TrainingStatusPowerMode = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  max: 'max',
} as const;

export interface TrainingStatus {
  status: TrainingStatusStatus;
  epoch: number;
  step: number;
  /** @nullable */
  loss?: number | null;
  /** @nullable */
  validationLoss?: number | null;
  /** @nullable */
  perplexity?: number | null;
  learningRate: number;
  tokensProcessed: number;
  tokensPerSecond: number;
  trainingTimeSeconds: number;
  powerMode: TrainingStatusPowerMode;
  activeWorkers: number;
  /** @nullable */
  startedAt?: string | null;
}

export type TrainingStartInputPowerMode = typeof TrainingStartInputPowerMode[keyof typeof TrainingStartInputPowerMode];


export const TrainingStartInputPowerMode = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  max: 'max',
} as const;

export interface TrainingStartInput {
  powerMode?: TrainingStartInputPowerMode;
}

export type PowerModeInputPowerMode = typeof PowerModeInputPowerMode[keyof typeof PowerModeInputPowerMode];


export const PowerModeInputPowerMode = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  max: 'max',
} as const;

export interface PowerModeInput {
  powerMode: PowerModeInputPowerMode;
}

export type WorkerStatus = typeof WorkerStatus[keyof typeof WorkerStatus];


export const WorkerStatus = {
  idle: 'idle',
  running: 'running',
  paused: 'paused',
  error: 'error',
} as const;

export interface Worker {
  id: number;
  name: string;
  type: string;
  status: WorkerStatus;
  queueSize: number;
  processed: number;
  errors: number;
  tokensPerSecond: number;
  /** @nullable */
  currentTask?: string | null;
}

export type DatasetStatus = typeof DatasetStatus[keyof typeof DatasetStatus];


export const DatasetStatus = {
  ready: 'ready',
  generating: 'generating',
  cleaning: 'cleaning',
} as const;

export interface Dataset {
  id: number;
  category: string;
  totalSamples: number;
  qualityScore: number;
  sizeKb: number;
  status: DatasetStatus;
  createdAt: string;
  updatedAt: string;
}

export type DatasetOverviewByCategoryItem = {
  category: string;
  count: number;
  quality: number;
};

export interface DatasetOverview {
  totalSamples: number;
  totalCategories: number;
  averageQuality: number;
  totalSizeKb: number;
  byCategory: DatasetOverviewByCategoryItem[];
}

export interface DatasetGenerateInput {
  category: string;
  count?: number;
}

export interface Checkpoint {
  id: number;
  name: string;
  epoch: number;
  step: number;
  loss: number;
  sizeMb: number;
  isActive: boolean;
  createdAt: string;
}

export type AgentStatus = typeof AgentStatus[keyof typeof AgentStatus];


export const AgentStatus = {
  active: 'active',
  idle: 'idle',
  thinking: 'thinking',
  error: 'error',
} as const;

export interface Agent {
  id: number;
  name: string;
  type: string;
  status: AgentStatus;
  lastAction: string;
  taskCount: number;
  createdAt: string;
}

export interface ModelInfo {
  name: string;
  version: string;
  parameters: number;
  layers: number;
  heads: number;
  dModel: number;
  dFf: number;
  vocabSize: number;
  maxSeqLen: number;
  architecture: string;
  activeParameters: number;
  memoryMb: number;
}

export type TokenizerStatsMostFrequentItem = {
  token: string;
  count: number;
};

export interface TokenizerStats {
  vocabSize: number;
  totalTokensSeen: number;
  uniqueTokens: number;
  mostFrequent: TokenizerStatsMostFrequentItem[];
  averageTokenLength: number;
}

export type HardwareMetricsRecommendedPowerMode = typeof HardwareMetricsRecommendedPowerMode[keyof typeof HardwareMetricsRecommendedPowerMode];


export const HardwareMetricsRecommendedPowerMode = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  max: 'max',
} as const;

export interface HardwareMetrics {
  cpuUsagePercent: number;
  ramUsedMb: number;
  ramTotalMb: number;
  gpuAvailable: boolean;
  /** @nullable */
  gpuUsagePercent?: number | null;
  /** @nullable */
  gpuVramUsedMb?: number | null;
  storageFreeMb: number;
  storageTotalMb: number;
  uptimeSeconds: number;
  recommendedPowerMode: HardwareMetricsRecommendedPowerMode;
}

export interface TrainingMetric {
  id: number;
  epoch: number;
  step: number;
  loss: number;
  /** @nullable */
  validationLoss?: number | null;
  perplexity: number;
  learningRate: number;
  tokensPerSecond: number;
  createdAt: string;
}

export type TrainingLogLevel = typeof TrainingLogLevel[keyof typeof TrainingLogLevel];


export const TrainingLogLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  success: 'success',
} as const;

export interface TrainingLog {
  id: number;
  level: TrainingLogLevel;
  message: string;
  /** @nullable */
  workerName?: string | null;
  createdAt: string;
}

export type GetTrainingMetricsParams = {
limit?: number;
};

export type GetTrainingLogsParams = {
limit?: number;
};