# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Babis M1 AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/sessions:
    get:
      operationId: listChatSessions
      tags: [chat]
      summary: List all chat sessions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSession"
    post:
      operationId: createChatSession
      tags: [chat]
      summary: Create a new chat session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
  /chat/sessions/{sessionId}:
    delete:
      operationId: deleteChatSession
      tags: [chat]
      summary: Delete a chat session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chat/sessions/{sessionId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages in a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and get streaming response
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: Streaming response (text/event-stream)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # ── Training ──────────────────────────────────────────────────────────────
...
  /training/start:
    post:
      operationId: startTraining
      tags: [training]
      summary: Start training
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingStartInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingStatus"
...
  /training/power-mode:
    post:
      operationId: setPowerMode
      tags: [training]
      summary: Change power mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PowerModeInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingStatus"
...
  /training/logs:
    get:
      operationId: getTrainingLogs
      tags: [training]
      summary: Get training logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingLog"

  # ── Workers ───────────────────────────────────────────────────────────────
...
  /checkpoints/{checkpointId}/load:
    post:
      operationId: loadCheckpoint
      tags: [checkpoints]
      summary: Load a checkpoint
      parameters:
        - name: checkpointId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingStatus"

  # ── Agents ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ChatSession:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, messageCount, createdAt, updatedAt]
    ChatSessionInput:
      type: object
      properties:
        title:
          type: string
    Message:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        thinkingMode:
          type: boolean
        createdAt:
          type: string
      required: [id, sessionId, role, content, createdAt]
    MessageInput:
      type: object
      properties:
        content:
          type: string
        thinkingMode:
          type: boolean
      required: [content]
    TrainingStatus:
      type: object
      properties:
        status:
          type: string
          enum: [idle, initializing, running, paused, stopped, error]
        epoch:
          type: integer
        step:
          type: integer
        loss:
          type: ["number", "null"]
        validationLoss:
          type: ["number", "null"]
        perplexity:
          type: ["number", "null"]
        learningRate:
          type: number
        tokensProcessed:
          type: integer
        tokensPerSecond:
          type: number
        trainingTimeSeconds:
          type: integer
        powerMode:
          type: string
          enum: [low, medium, high, max]
        activeWorkers:
          type: integer
        startedAt:
          type: ["string", "null"]
      required:
        [
          status,
          epoch,
          step,
          learningRate,
          tokensProcessed,
          tokensPerSecond,
          trainingTimeSeconds,
          powerMode,
          activeWorkers,
        ]
    TrainingStartInput:
      type: object
      properties:
        powerMode:
          type: string
          enum: [low, medium, high, max]
    PowerModeInput:
      type: object
      properties:
        powerMode:
          type: string
          enum: [low, medium, high, max]
      required: [powerMode]
...
    TrainingLog:
      type: object
      properties:
        id:
          type: integer
        level:
          type: string
          enum: [info, warn, error, success]
        message:
          type: string
        workerName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, level, message, createdAt]